test(infra): #4199 Node 版 fitness の手書き列挙を glob 化し workflow 直書き回帰を検出する
